requirement.txt
Django==1.7.7
Fabric==1.4.3
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==2.8.1
South==1.0.2
Werkzeug==0.10.4
coverage==3.7.1
django-bootstrap3==5.1.1
django-extensions==1.5.2
django-registration-redux==1.1
djangorestframework==3.1.1
gnureadline==6.3.3
googlemaps==2.2
gunicorn==19.3.0
itsdangerous==0.24
logilab-common==0.63.2
pycrypto==2.6
selenium==2.45.0
six==1.9.0
ssh==1.7.14
wsgiref==0.1.2